Adding accuracy to generic parametric geometry intersection tests
